reorg code - reorganized code such that the work for info in in a single file, and the work for github is in another
